tikleap_docker/client/requirements.txt
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.10
loguru==0.7.3
lxml==6.0.0
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
rnet==2.3.7
rsa==4.9.1
schedule==1.2.2
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.5.0
